feat(examples): improve composition example with dlog or dleq to have same level of details as simple example
